+++ R/date.R
@@ -12,13 +12,10 @@
 #' @examples
 #' obj_date <- date("dates.xlsx", id, dplyr::contains)
 #' obj_date <- date("dates.xlsx", id, dplyr::contains, "uneq")
-#'
 date <- function(file, id, get_visit, get_date = dplyr::contains, str_date = "DAT") {
-  id <- dplyr::enquo(id)
-
   obj <- list(
     file = file,
-    id = id,
+    id = dplyr::enquo(id),
     get_visit = get_visit,
     get_date = get_date,
     str_date = str_date
@@ -50,45 +47,44 @@ date <- function(file, id, get_visit, get_date = dplyr::contains, str_date = "DA
 #' bio_date_E3 <- c("1991-03-23", "1991-03-16", "1991-03-16")
 #'
 #' df <- data.frame(id, screen_date_E1, rand_date_E2, ph_date_E3, bio_date_E3,
-#' stringsAsFactors = FALSE)
+#'   stringsAsFactors = FALSE
+#' )
 #'
 #' timeline <- system.file("dates.xlsx", package = "dmtools")
 #' obj_date <- date(timeline, id, dplyr::contains)
 #'
 #' obj_date <- check(obj_date, df)
 #' choose_test(obj_date, "out")
-#'
 choose_test.date <- function(obj, test = "out", group_id = T) {
-
   result <- obj %>% get_result(group_id)
 
   # filter final dataset
   if (test == "out") {
     result <-
       result %>%
-      dplyr::filter(.data$is_in_timeline == F) %>%
-      dplyr::select(-.data$is_in_timeline, -.data$is_equal, -.data$stand_equal)
+      dplyr::filter(.data$IS_TIMELINE == F) %>%
+      dplyr::select(-.data$IS_TIMELINE, -.data$IS_EQUAL, -.data$EQUALDAT)
   } else if (test == "uneq") {
     result <-
       result %>%
-      dplyr::filter(.data$is_equal == F) %>%
+      dplyr::filter(.data$IS_EQUAL == F) %>%
       dplyr::select(
-        -.data$is_equal,
-        -.data$standard_interval,
-        -.data$standard_date,
-        -.data$standard_name,
-        -.data$out
+        -.data$IS_EQUAL,
+        -.data$PLANDAT,
+        -.data$STARTDAT,
+        -.data$STARTVISIT,
+        -.data$DAYS_OUT
       )
   } else if (test == "ok") {
     result <-
       result %>%
-      dplyr::filter(.data$is_equal == T & .data$is_in_timeline == T) %>%
-      dplyr::select(-.data$is_in_timeline, -.data$is_equal, -.data$out)
+      dplyr::filter(.data$IS_EQUAL == T & .data$IS_TIMELINE == T) %>%
+      dplyr::select(-.data$IS_TIMELINE, -.data$IS_EQUAL, -.data$DAYS_OUT)
   } else if (test == "skip") {
     result <-
       result %>%
-      dplyr::filter(is.na(.data$date_item)) %>%
-      dplyr::select(-.data$out, -.data$standard_interval, -.data$is_in_timeline, -.data$is_equal, -.data$stand_equal)
+      dplyr::filter(is.na(.data$VISDAT)) %>%
+      dplyr::select(-.data$DAYS_OUT, -.data$PLANDAT, -.data$IS_TIMELINE, -.data$IS_EQUAL, -.data$EQUALDAT)
   } else {
     stop("uknown parameter ", test)
   }
@@ -105,7 +101,7 @@ choose_test.date <- function(obj, test = "out", group_id = T) {
 #' @return A data frame. Visit's dates.
 #'
 find_colnames.date <- function(obj, dataset, row_file) {
-  str_visit <- row_file$num_visit
+  str_visit <- row_file$VISITNUM
   string_date <- obj[["str_date"]]
   get_visit <- obj[["get_visit"]]
   get_date <- obj[["get_date"]]
@@ -139,59 +135,58 @@ run_tests.date <- function(obj, dataset, row_file, date) {
   id <- obj[["id"]]
 
   # params of visit
-  minus <- as.integer(row_file$minus)
-  plus <- as.integer(row_file$plus)
-  shift <- as.integer(row_file$shift)
-  name_visit <- row_file$name_visit
-  st_date <- row_file$standard_date
-  st_name <- row_file$standard_name
+  minus <- as.integer(row_file$MINUS)
+  plus <- as.integer(row_file$PLUS)
+  shift <- as.integer(row_file$VISITDY)
+  name_visit <- row_file$VISIT
+  st_date <- row_file$STARTDAT
+  st_name <- row_file$STARTVISIT
 
   # params for check if dates are equal
-  check_equal <- as.logical(row_file$check_equal)
-  equal_date <- ifelse(check_equal, row_file$equal_date, date)
+  check_equal <- as.logical(row_file$IS_EQUAL)
+  equal_date <- ifelse(check_equal, row_file$EQUALDAT, date)
 
   # check dates
   dataset %>%
     dplyr::mutate(
-      standard_name = st_name,
-      standard_date = lubridate::as_date(.data[[st_date]]),
-      name_event = name_visit,
-      name_item = date,
-      date_item = lubridate::as_date(.data[[date]]),
-      stand_equal = lubridate::as_date(.data[[equal_date]])
+      STARTVISIT = st_name,
+      STARTDAT = lubridate::as_date(.data[[st_date]]),
+      VISIT = name_visit,
+      TERM = date,
+      VISDAT = lubridate::as_date(.data[[date]]),
+      EQUALDAT = lubridate::as_date(.data[[equal_date]])
     ) %>%
-    dplyr::mutate(standard_interval = lubridate::interval(
-      .data$standard_date - lubridate::days(minus),
-      .data$standard_date + lubridate::days(plus)
+    dplyr::mutate(PLANDAT = lubridate::interval(
+      .data$STARTDAT - lubridate::days(minus),
+      .data$STARTDAT + lubridate::days(plus)
     ) %>% lubridate::int_shift(lubridate::duration(days = shift))) %>%
-    dplyr::mutate(is_in_timeline = .data$date_item %within% .data$standard_interval) %>%
-    dplyr::mutate(is_equal = .data$date_item == .data$stand_equal) %>%
-    dplyr::mutate(out = ifelse(.data$is_in_timeline, 0, calc_diff(.data$standard_interval, .data$date_item))) %>%
+    dplyr::mutate(IS_TIMELINE = .data$VISDAT %within% .data$PLANDAT) %>%
+    dplyr::mutate(IS_EQUAL = .data$VISDAT == .data$EQUALDAT) %>%
+    dplyr::mutate(DAYS_OUT = ifelse(.data$IS_TIMELINE, 0, calc_diff(.data$PLANDAT, .data$VISDAT))) %>%
     dplyr::select(
       !!id,
-      .data$standard_name,
-      .data$standard_date,
-      .data$name_event,
-      .data$name_item,
-      .data$date_item,
-      .data$standard_interval,
-      .data$stand_equal,
-      .data$is_in_timeline,
-      .data$is_equal,
-      .data$out
+      .data$STARTVISIT,
+      .data$STARTDAT,
+      .data$VISIT,
+      .data$TERM,
+      .data$VISDAT,
+      .data$PLANDAT,
+      .data$EQUALDAT,
+      .data$IS_TIMELINE,
+      .data$IS_EQUAL,
+      .data$DAYS_OUT
     )
 }
 
 #' Function for calculating the difference between two dates
 #'
-#' @param st_inter An interval. An Object of interval.
-#' @param dt_item A date item. An Object of date.
+#' @param st_inter An interval. An object of interval.
+#' @param dt_item A date item. An object of date.
 #'
 #' @return An integer scalar. Differences between the two dates.
 #'
 calc_diff <- function(st_inter, dt_item) {
   start_item <- dt_item - lubridate::as_date(lubridate::int_start(st_inter))
   end_item <- dt_item - lubridate::as_date(lubridate::int_end(st_inter))
-  minimum <- pmin(abs(start_item), abs(end_item))
-  ifelse(minimum == abs(start_item), paste0(minimum, "<-"), paste0("->", minimum))
+  pmin(abs(start_item), abs(end_item))
 }