+++ R/dmtools.R
@@ -1,7 +1,7 @@
 #' dmtools: package to validate data
 #'
 #' for checking the dataset from EDC in clinical trials
-#' 
+#'
 #' @name dmtools
 #' @docType package
 #' @importFrom magrittr %>%