+++ tests/testthat/test_date.R
@@ -18,13 +18,13 @@ test_that("check the number of dates, which is out the timeline", {
 })
 
 test_that("check names of dates, which is out the timeline", {
-  expect_equal(out_date$name_item[1], "bio_date_E3")
-  expect_equal(out_date$name_item[2], "screen_date_E1")
+  expect_equal(out_date$TERM[1], "bio_date_E3")
+  expect_equal(out_date$TERM[2], "screen_date_E1")
 })
 
 test_that("check the difference of protocol", {
-  expect_equal(out_date$out[1], "->2")
-  expect_equal(out_date$out[2], "1<-")
+  expect_equal(out_date$DAYS_OUT[1], 2)
+  expect_equal(out_date$DAYS_OUT[2], 1)
 })
 
 test_that("check the number of dates, which is no equal", {